feat(block): Make block mergable
A factory typically creates one specific operation (Block) to be used in
a sequence that require multiple blocks. We want to be able to use the
editing capabilities on the block (tranlate, scale etc) to manipulate
the "group" of blocks that we have created. With this, we can merge
blocks together and run operations on the union.

Note: This could be made using a BlockGroup class, that takes in blocks,
and has the same editing capabilities as the blocks have. However, this
will lead to more complexity at this moment.