refactor(package): Prepare for distribution
TypeScript doesn't have support for rewriting import paths, even though
they have support for baseUrl and paths compiler option. We therefore
must alter the import paths after tsc has built the code. This is done
using 'tsc-alias'.

https://github.com/microsoft/TypeScript/issues/15479

For the test case, we use 'tsconfig-paths' package to solve this in
runtime (because we don't recompile the entire project on each save when
watching tests).